+++ apix-integrationtests/60/src/main/java/eu/xenit/apix/Dummy.java
@@ -1,5 +1,0 @@
-package eu.xenit.apix;
-
-public class Dummy {
-
-}